+++ src/types.d.ts
@@ -27,6 +27,8 @@ export interface Options {
   changedFiles?: ChangedFiles | null
   multipleFiles?: string[]
   multipleJunitFiles?: string[]
+  showFailedTests?: boolean
+  maxFailedTests?: number
 }
 
 export interface ChangedFiles {
@@ -54,11 +56,7 @@ export interface Summary {
 }
 
 export type CoverageColor =
-  | 'red'
-  | 'orange'
-  | 'yellow'
-  | 'green'
-  | 'brightgreen'
+  'red' | 'orange' | 'yellow' | 'green' | 'brightgreen'
 
 export interface SummaryReport {
   summaryHtml: string
@@ -72,10 +70,20 @@ export interface Junit {
   failures: number
   errors: number
   time: number
+  failedTests: FailedTest[] // calculated field
+}
+
+export interface FailedTest {
+  suiteName: string
+  testName: string
+  message: string
+  file?: string
+  line?: number
 }
 
 export interface JunitReport extends Junit {
   junitHtml: string
+  failedTestsHtml: string
 }
 
 export interface CoverageLine {